Added file check before transcribing #292
When looping through files, if a file is not valid instead of showing error message right away and preventing other files from being trancribed, it would add it to a count and process next files in the list.
Once completed either success message if all files are transcribed or error message with number of files that failed transcription.